Mesa-19.0.1.
Now using a release build.
Drivers are limited to those which can be built on linux in the X86 architecture,
but all available drivers, even for old radeon versions, are again listed.

Note that tests now exist, and also that reducing the number of drivers will not save much time or space.

git-svn-id: svn://svn.linuxfromscratch.org/BLFS/trunk/BOOK@21425 af4574ff-66df-0310-9fd7-8a98e5e911e0