Various rewording in rustc:
Primarily, do not indicate that a few tests should be ignored.
Without gdb, or with gdb-9.2, only two testsreliably failed.
But with gdb-10.1, which is newer than this version of rust,
four extra failures have appeared because of changes in gdb.
We hold on to versions of rust, but we update gdb when that is
issued, so it semes likely that the number of failures from
newer gdb may increase. Ratherthan play whack-a-mole with
extra seds as tests are reproted to now fail, let them fail.

git-svn-id: svn://svn.linuxfromscratch.org/BLFS/trunk/BOOK@23857 af4574ff-66df-0310-9fd7-8a98e5e911e0