Firefox security updates:
Firefox-91.0esr
JS78 using firefox-78.13.0esr
and introducing firefox-78.13.0esr as Firefox Legacy.

The first two have been tested and measured on glibc-2.34 systems.
The latter has been measured on a slightly older system where 91.0
fails to build.

If anyone likes ff78 so much that they want to use it on a
glibc-2.34 system, feel free to create the necessary patches
but note that I expect to remove Firefox Legacy in November,
it is only a transitional item.