kf5-framworks: make kapidox dependencies more precise
Some python modules are in the book, but would be downloaded if not
present at build time. Some other python modules are not in the book,
but are downloaded at build time. Make python modules present in the
book recommended (and not tun time), and add those not present in the book.