Mesa - add comment for editors, fix wiki link.
The crocus driver will be automatically selected at some point,
the fix for that is already in master.

I needed to add a comment to the wiki for using crocus, which is
not yet the default - the link was broken, he old MesaLib page
is the right place, but last updated 15 years ago and the existing
details seem anachronistic.