mesa - do not recommend virgl for qemu -
It needs virgl-renderer, and although qemu linked to that works
fine in certain places, such as reducing the cpu overhead from
large-window glxgears when running in a BLFS-style qemu build,
offloading the guest's rendering to the host GPU appears to
need libvirt and a much fuller qemu build.  It seems to be mainly
for gamers.

For information, link to virglrenderer - I was not logged in to
freedesktop when I tried that today, so note the need for an
account there but also link to the current rlease for those who
do not have an account.

In passing, note that both Fedora and AUR appear to be using later
git commits than the 0.9.1 release.