resurrect ntfs-3g, and mention the new in-kernel NTFS impl
The tools (like mkfs.ntfs) is still useful, and one may be puzzled if he
uses:

    mount /dev/sdb1 /mnt

Then `mount` will find out that /dev/sdb1 is NTFS, and use the *old*
NTFS support (read-only) in kernel for it.  The user will then be
puzzled. Mention how to create a wrapper script for `mount` to use the
new NTFS3 driver for ntfs.