Fix PAM config files of shadow apps
shadow applications chpasswd and newusers use the "password" type,
and expect to be able to pass the password to the PAM module. But
we use pam_permit.so, which does nothing except return PAM_SUCCESS.
So the applications themselves do nothing without returning an
error. Change the config files to include system-password.
Also clean up the config files so that only the types used by the
applications appear.
Fixes #15950