Move Grantlee from General Libraries to X Libraries.
I've had a TODO list item after a Discussion with Bruce about this particular
package since last package freeze. Because of it's reliance on Qt5 and the fact
that it's only used for KHelpCenter, it really does not belong in
General Libraries. Moving it to X Libraries should suffice in this case,
since it doesn't really fit in KF5 either.

In addition, update the dependencies and the installed files list, and
add short descriptions for the two libraries installed by this package.