Qt5 - apply kf5.15 patches from kde.
I have only used the patches for modules used by packages in BLFS
or required by the extra KDE packages mentioned in BLFS. For the
full details of what has been applied (or omitted) see
https://www.linuxfromscratch.org/~ken/qt5/

I have also revised the SBU - the extra patches make a minimal
difference on my -j4 builds, but I need to use -j8 to get to 22
SBU. That figure seems to have been carried over from when we were
using gcc-10.2.

In the optional dependencies, pciutils was shown as optional but
required for qtwebengine: show it as optional, other parts of qt
can use it, and add it to qtwebengine as required.  Remove python2
from the qt5 optional dependencies, only qtwebengine needs it and
it is already shown as required there.