polkit: use archive tarball and enable tests
Normally we perfer release tarballs than archives.  But for polkit, we
are using meson so the generated configure script is not needed.  And,
the release tarball lacks test support files and prevents us from
running tests.

For such a "security related" package, skipping test seems not good...
(That being said, we'd been busying fix CVEs not found by the test. :( )