gst-plugins-bad: add libva as optional
There is a new gstreamer vaapi plugin in gst-plugins-bad, which seems
designed for replacing gstreamer-vaapi.  But it's still experimental now
and not enabled unless it's set in GST_PLUGIN_FEATURE_RANK explicitly.

I've used this plugin for a while because gstreamer-vaapi does not work
fine on my system with iHD_drv_video.so.

See https://gitlab.freedesktop.org/gstreamer/gstreamer-vaapi/-/issues/312.