Update to k3b-22.04.0.
This is an out-of-cycle update to k3b.  This version does not use ffmpeg
libraries.  The previous version does use those libraries but the ffmpeg-5
libraries do not work for that version.  This version of k3b optionally uses
ffmpeg as an external executable which does not conflict with this package.