Fixes for gcc-12
for qtwebengine and xf86-video-vmware. For qtwebengine, I am
not sure it is not upstream, but what I have added works:
the problem is with the definition of functions std::begin and
std::end. They are defined in several headers (among which
<vector>), and in a different way in some other headers. Looks
like the former headers were included with gcc-11 and below,
while they are not anymore with gcc-12 (there is no standard
on what is included by what in C++ headers). The problem with
qtwebengine (or rather the third party skia module) is that
they also have towers of headers including each other, so that
it is hard to find where the missing C++ header has been included,
if it has been.