gnome-keyring: switch to Git snapshot for 42.0
The maintainer did not pay enough attention using "make dist" to
generate the release tarball.  See:
https://gitlab.gnome.org/GNOME/gnome-keyring/-/issues/109

bdubbs has partially worked around the issue, but on systemd we still
get:

    make[2]: *** No rule to make target
    'daemon/gnome-keyring-daemon.socket', needed by 'all-am'.  Stop.

Instead of fetching this .socket file separately, it's easier to just
download a clean tarball and regenrate the configure script ourselves.
I'd prefer them to switch to meson and avoid this kind of problem
forever.