Add libsoup-3.0.6.
libsoup3 should be OK now.  There have not been any issues discovered
as of the date of this commit.

gvfs is also changed to add back http.  Also avahi is now recommended.
avahi can be removed by adding -Ddnssd=false but that requires a patch.

  https://gitlab.gnome.org/GNOME/gvfs/-/issues/621

This has been fixed upstream.