add libsoup3 as optional for tracker3 and gst10-plugins-good
Both packages handle libsoup2 and libsoup3 as follows: if one libsoup is
already loaded by the process, it will use the loaded one.  Otherwise,
try to load libsoup3 first, and if it's not found try libsoup2.

But they need the headers of one libsoup version at build time to
support loading this version at runtime, so we can't make the libsoups
runtime dependency.

For tracker 3, also demote libsoup2 to recommended.  It can be built with
libsoup3 only (and the building system will do it silently if libsoup3
is installed but libsoup2 not).  But I'm not sure if it will cause
runtime issues.