firmware: stop using late loading
The kernel dev describes late loading as "just lottery and broken".
From Linux 5.19 it will produce a warning and taint the kernel.  And, we
can simply compare the microcode version number to know if there is an
update so there is no reason to use late loading at all.

Update the example for Intel microcode by the way.