zsh: enable libcap and gdbm
They are already in LFS so there is no significant reason not to use
them.  Using gdbm may cause license issue but AFAIK as long as we don't
edit gdbm or zsh code in a hidden way we'll not violate the license.

And zsh does not use pam_cap (or any PAM module) at all.  Perhaps we
should perform a cleanup on pam_cap dependencies.