pipewire: add WirePlumber an as external optional dependency
Currently we only build pipewire for satisfying GNOME building
dependency.  I've switched to use Pipewire instead of Pulseaudio as the
audio daemon on my system.  For the Pipewire daemon to function
properly, a Pipewire session manager is needed.  The upstream recommends
to use WirePlumber.

It's too early to provide instruction for running a Pipewire daemon in
the book, but we can list the external dependency as it does no harm.