GDM: Add text on how to start GNOME from the command line.
This is useful for users who don't want a display manager, as well as
users who have the proprietary NVIDIA driver installed and want Wayland
support (since GDM won't show the session).

I placed this in the GDM page since GDM is still required for such a
session to work anyway (the .GIR files in particular), but I'm open to
moving it in the future if desired.