Fix building dvisvgm with ghostscript-10.0.0
In this version of ghostscript, some private symbols have been
hidden. One of the private symbols is gs_error_names, which is used
by dvisvgm. Fortunately, dvisvgm, has already the infrastructure
for this case, since private symbols are always hidden in Windows.
Upstream has a patch:
https://github.com/mgieseki/dvisvgm/commit/9bf81fd0
which can be reduced to a sed.