Python3's module reinstallation instructions:
Add meson into the list for both init systems

On systemd, mention MarkupSafe and Jinja2 from LFS as well.

There is a really long list of affected packages in IRC. We should come
up with a better way of handling this, I feel like the list would be too
long if we added it in normally.