Add pythonhosted page
This page is not complete.  The build part does not work yet, but
the fetch of files does.  If the fetch is run a second time, it
will only get a file in the .md5 file that is not already present.

The install instructs are mostly OK, but the test for the module
already installed is not correct yet.  The next commit will
update it.