mutter: use --buildtype=debugoptimized for now ...
instead of --buildtype=release to work around a crash in gnome-shell.

Upstream issue: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/5512

If close an focused app window via the Activities overview, an assertion
fail will be triggered in clutter_actor_get_preferred_{width,height}.
When debug is disabled, the assertions are disabled as well and the
issue turns into a NULL dereference, leading to a crash.

This is definitely not a proper fix.  But the issue has been haunting
gnome-42 and 43, and the upstream has not given any response yet.

Note that it's not a "compiler misoptimization" issue.  The assertion
fails even with -O0, i. e. if both the assertions and compiler
optimization are disabled, gnome-shell will crash as well.