Update to at-spi2-core-2.46.0
This update brings some relatively major changes which will be addressed
in subsequent commits. Please see ticket #17060 for more details.

In particular, at-spi2-atk and ATK need to be archived since this
package now provides both of them.

Dependencies need to be updated to account for this around the book.

I've build tested all of the packages that depend on this one which are
listed in the book.

The currency scripts will need to be updated. I'll try to take care of
this