Fix qt5 so that all headers are installed
The patch for qt-5.15.8 creates a header file, which needs to be
known by the build machinery. In a git repository, this is done
automatically, but in a build tree coming from a tarball, the header
file is ignored.
So we add a .git directory into qtbase, which fakes a git repository,
and allows the header file to be used and installed by the build system
(it may be needed when compiling programs using Qt thread mutexes).