Remove pythonhosted for now.
We will change pythonhosted.xml to incorporate all python modules
(perhaps excepting PyXDG) at https://files.pythonhosted.org
and move it to the top of the Python Modules page.

We will need to place the md5sums for the module sources in packages.ent
so they can be referenced in two places.

For now, remove the page from the build.  Later we will need to clean up
Makefile and archive stylesheets/{pyhosted-inc,pythonhosted}.xsl.