Hatch-vcs: use underscore in written text
It is not easy to make one's mind whether a dash or an underscore
should be used: pypi, and doc, use hatch-vcs. But tarball name
nd installed module directories have hatch_vcs (note that dashes
are forbidden in python module names). So use hatch_vcs in written
text (don't change the dash in the entities and the id).