Use mkdir -pv /opt/rustc-<version>
The most likely reason to rerun a scripted build of rust is that
something failed during build. But since this directory is created at
the beginning of the script, the script will fail early cresting the
directory.

Of course, now, there is no protection against erasing a successful
build when rerunning with different options.