xwayland: Add --buildtype=release, and ...
Adjust the description of man page Xserver.1: technically both Xorg and
Xwayland are in the category of "X server".  The reason we don't install
it here is not "it's not a part of this package", but we don't want to
overwrite a man page from another package.

Remove the creation of /etc/X11/xorg.conf.d: "Xorg" is "another X
server" so technically Xwayland (and other packages except Xorg and the
Xorg drivers) should have nothing to do with it.  In my last build
/etc/X11/xorg.conf.d is empty.