fftw: Add --enable-avx2, and document --enable-{sse2,avx,avx2,avx512}
The first AVX2-capable CPUs are released in 2013 (ten years ago)
and now all recent Intel Core and AMD Ryzen models support it.  So
it makes sense to enable it now.  AVX512F is much uncommon now for
non-server CPU models (only Rocket Lake, Alder Lake with E-cores
disabled, and Zen 4), so leave it as <option>.