libcdio: Document a test failure
This is a dead project, but the comment in the code claims the
cdio_realpath function is designed to emulate POSIX realpath while the
test case is not using POSIX-compatible realpath correctly.  So it seems
a test case issue, only document it instead of fix.

Note that this test has started to fail since Glibc-2.36, not 2.37.  But
somehow we didn't catch it before.