Attempt to add CacheControl and its dep.
CacheControl is used an inkscape clipart file, and in an inkman
file for searching for external packages and getting meta data
about them.

Unfortunately, I cannot find a working link.  When I go to
https://pypi.org/project/CacheControl/#files and click on the
link for CacheControl-0.12.11.tar.gz I get a good download,
but if I mouse over that to read where it points to I get
a very long link which starts
 https://files.pythonhosted.org/packages/49/9b/
followed by \n unfeasibly long hash which only remains visible
while my mouse is on the link.

Trying to point to c/CacheControl fails.

Current upstream has v0.12.12 but there is at least one issue for
why that might not be at pythonhosted, and trying to link to the
tag for 0.12.11 (there are no releases there) fails.