grub-setup: Adjust kernel config
With CONFIG_EXPERT disabled in LFS, CONFIG_FRAMEBUFFER_CONSOLE should be
forced to y.  Update the menu to reflect this.

CONFIG_NLS is also forced to y with CONFIG_ACPI and CONFIG_PCI enabled
(both should be necessary for any x86 system produced in 21st century).

Don't encourage user to be too "advanced".