grub-efi: Fix "blind mode" issue
Move "insmod all_video" out of the "if loadfont" statement.  It's enough
for the kernel to use EFI FB.

The "terminal_output gfxterm" statement has to be guarded by "if
loadfont", or it can garble the display if unicode.pf2 not available.

"gfxmode" seems no-use on EFI.  According to upstream doc it's only
used for VGA BIOS Extension (as the name suggests, not available in EFI
environment).