colord: Make systemd dependency runtime, and add elogind dependency
It uses logind API in libsystemd.so/libelogind.so.  Not sure if it's
strong enough to be "recommended", we may demote it later.

Q: Why systemd dependency is runtime, but elogind not?
A: LFS systemd provides libsystemd.so, which contains the logind APIs.
But these APIs will fail at runtime because systemd-logind is missing in
LFS, so we need a runtime dependency.  LFS sysv does not provide
libsystemd.so, and libsystemd.so is a symlink to libelogind.so which is
a part of elogind, so it's a build time dependency.