evince: Take the advantage of /usr/lib/libkpathsea.so symlink
So we can simply the instruction.  Also remove a stale command
explanation referring to "configure script" (we are using meson now).

Simply the explanation for libkpathsea.so symlink in texlive too as we
don't distinguish "old" and "new" evince installations anymore.  The
original explanation emphasizing "old" evince installation also does not
make sense to me: we have /etc/ld.so.conf addition which should be
enough for ld.so to find libkpathsea.so.6 when evince is used; and ld.so
only searches for "libkpathsea.so.6" (which is a DT_NEEDED entry in
evince libdvidocument.so), not unversioned libkpathsea.so.