Update to LLVM-16.0.3 (#17861)
With Thunderbird workaround applied, we can try to update to LLVM 16
again.  Note that Firefox may need a similar workaround but at least it
does not crash on start-up.

This reverts commit 218e13ce4ba083720efc136aedb40c5fa0d9eaf2
("Revert back to LLVM-15.0.7 for now"), bump LLVM to 16.0.3, and updated
md5sums for tarballs and test suite status for it.