rust: Add rustdoc into build.tools
The rustdoc program is a part of rustup default install.  And it was
installed with previous rustc releases even if we didn't specify it in
build.tools (so it's in the "contents and short descriptions" section).
But with rustc-1.69 it's not installed if we don't explicitly add it
into build.tools.

It's needed by "cargo doc" command (I use it when I develop Rust
crates), and also some BLFS packages (at least, the testsuite of
librsvg).

Add rustdoc into build.tools so it will be built and installed.