Move Vulkan description from "Graphical Environments" to Mesa
Well, Vulkan is not an alternative of X, nor Wayland.  It's an
alternative of OpenGL.

Change "libvulkan" to "Vulkan-Loader" by the way.  "libvulkan" is a
library, and Vulkan-Loader is the name of the package providing it.  The
URL is also updated because there is no download link at vulkan.org.