elogind: Fix the desc for -Dcgroup-controller
The value of this switch only makes a difference with cgroupv1.  But we
are using Linux 6.x, and elogind is built with
-Ddefault_hierarchy=unified (the default), so it actually uses cgroupv2
and it seems we don't really need this switch.

But after a closer look at the building system I realized if the kernel
is not built with CONFIG_CGROUPS enabled (for example in a chroot
environment where the host does not enable this), the building system
will refuse to build the package w/o an explicit setting here.  So
update the description.