js102: Fix building on i686 again.
The default build configuration in 102.13 appears to have moved back to
GCC. Rather than change it, let's just modify math_private.h again with
a simpler tweak to allow it to build properly again (wrapped in a case
statement).

This particular problem resurfacing is quite odd. I can confirm that
js-102.12.0 does work though, but updated due to the JS CVE.

Firefox 115.x works well on i686 and builds without any tweaks, so
whenever we move to js115, this should not be a problem anymore.