Add note about future removal of context.
This only applies to texlive source, at least for TL2023 the
x86_64 binaries include luametatex.  For source the luametatex
sources are not included. In the unlikely case that we have a
user who both builds texlive from source and uses context, let
them know before we release 12.0 in case another luatex
vulnerability is discovered before TL2024.