Drop multicast and broadcast packets in iptables
The reason is that our own box can output those kind of packets
rather frequently, and those packets immediately appear as
incoming. Since any non expected packet is logged, this clutters
the log. Dropping them in the "personal firewall" case is what is
wanted, but it has to be done before logging.