Add a section to xorg-config about issues with Nouveau GPUs.
If we had the Xorg Drivers section still with xf86-video-nouveau there,
this probably would've gone over there originally, but this place fits
well.

The original plan was to put this into LFS, but I decided against it
since it works fine until you start using programs such as Epiphany or
KDE Plasma.