Revert the additional instructions for Crocus and GDM.
I had the Intel Media Driver installed on this system at some point in
the distant past, and it installed /etc/profile.d/intel-media.sh - which
overrode the libva-intel-driver, and caused gstvaapi.so to fail to load
correctly on Wayland... which then caused GDM to fail to start. Note
that X11 worked fine because of using a different code path.