building-notes: Note how to use cgroup for limiting resource usage
We were saying "-jN means using N cores (or N threads)".  This is
completely wrong. "-jN" only tells the building system to run N jobs
simultaneously, but each job can start their own subprocesses or threads
and there is no way for the building system to know how many
subprocesses or threads a job will start.

This caused a lot of misunderstandings and encouraged users to wrongly
blame building systems.

Fix the description of -jN, and add how to use cgroup to control the
usage of CPU cores and system RAM.

On a systemd-based system, systemd is the cgroup manager and manually
operating on cgroups may puzzle systemd.  So use systemd-run for
creating and setting up cgroup.  On a sysv-based system create and set
up the cgroup manually.