appstream: Add /usr/share/metainfo/org.linuxfromscratch.lfs.xml file
It fixes the "Operating System Details" content in appstreamcli sysinfo
output.

Note that the file name is derived from HOME_URL and ID properties from
/etc/os-release, and the HOME_URL property is just added (at LFS
r12.0-44).  If HOME_URL is not set, the file should be renamed to
"lfs.xml" and the <id> in the file should be "lfs", instead of
"org.linuxfromscratch.lfs".