mesa: Use a sed instead of patch for LLVM-17 fix
When the patch can be converted to a not-so-long sed, we prefer the sed
because it tells people "what this command is doing" more explicitly and
also reduces an additional download.  And for patch or sed we need a
<para> describing "what it fixes".

It's also a bad idea to fold the patch/sed command just before "meson"
in the same <screen> because it'd be too easy to overread it.