Make Qt5 recommended for poppler
Recently, upstream has changed the way Qt5 is detected: if Qt5
is not installed, you have to pass -DENABLE_QT5=OFF, otherwise
cmake stops with an error. Previously, a non present Qt5 was
not fatal. But passing -DENABLE_QT5=OFF when qt5 is installed
prevents building the poppler-qt5 library, which is needed by
okular, one of the main users of poppler.
Until now we had Qt5 as optional, since building full Qt5 for other
users of poppler (mainly inkscape, other users can benefit of qt5)
was kind of overkill. But now we have qt5-alternate, and also
this change in behavior of poppler. So it seems it is acceptable
to have qt5 or qt5-alternate as recommended.