Fix and simplify 'Prefer Chosen CJK fonts' :
I had a setup like this on one of my machines, now that I'm
looking at the detail of fontconfig in a local branch I discovered
that there were certain problems with the example:

1. I'm in an en locale, for pages that do not specify a locale
(or in vim/view, e.g. in mutt) the Japanese fonts were being
preferred.

2. Fontconfig does not consider UMing suitable for zh-sc so it
was hardly ever used - and it does not really belong in local.conf.

3, Really prefer a Japanese font for Sans Serif and monospace, but
no point listing two of them.

3. Comment where WenQuanYi Zen Hei is regarded as adequate and
therefore do not include it in these preferences, since it will
be picked up after them.