Add qt6
This is a rough version of the new Qt6 page.  The installation
instructions should be OK, but the Dependencies, Contents, and
Short Descriptions need to be checked.

The instructions have not yet been checked on a systemd system.

There is no ChangeLog entry yet. It will be added when the page
has been validated.