tuningfonts:
In talking about hinting, mention dots per inch - some people can
detect colour fringing if the actual DPI is a little different from
96 dpi, and general recommentations for High DPI screens are to
disable hinting because it is not required when the font is increased
in size to have the expected size, i.e. more pixels are used for
the glyph so they can be either off or on rather than shades of grey.