tuningfonts: Pango no-longer supports full hinting.
Details eventually found (from 2020), with a very ugly set of bug
responses.

Note: Putting the commented link to the bug on a new line adds a
blank line to the rendered file, and there is already a blank line.
Moving the comment to the same line as the text fixes that.