tuningfonts -
Italic is a style, not a face - make the fontconfg page match
the TTF-and-OTF-fonts page.

Do not use caps for 'variable', it is not a proper noun.

Typos.

Change one or two 'normal' to other words: too many uses of normal
throughout the page. Mozilla's CSS doc describes weight 400 as
'Normal (Regular)', so for weights normal is the correct word
(Regular only applies to non-book non-oblique non-slant styles).