tuningfonts:
Rework second fc-match example:

'fcmatch -a TYPE | less' is useless: -a functions similarly to
fc-list, any following parameter is ignored. So the example listed
all installed font-style-weight combinations for all three TYPEs.

Replace by plain fc-match TYPE.

Remove TYPE specification from introductory paragraph, it is only
used for the replacement command here. Use the same markup for the
command and in the explanations, so that a capital TYPE in the
explanations is replaced by a smaller slanted TYPE.

The command after this shows a fuller variant with the type, lang
and weight being specified.