vim: Make runtime update command nodump
Well, we generally do not like an update of "a part of the package",
esp. in this case we are doing an unversioned update.  This makes the
book unstable, i.e. if a user executes this command a month after the
release of the book (s)he may get different results.  And I don't think
it's a good idea to make vim &build-use-internet;.

And when we update vim to a new version (from 9.0.2103 to 9.0.2189 for
example), "make install" would already update the runtime files to
9.0.2189.

So just make the runtime update command nodump and move rsync to
optional.