seamonkey: Document sudo/su may kill $PATH_PY311
Like rustc.  Move the unset command out of the <screen> for
installation because if we use sudo --preserve-env=PATH_PY311 -s for a
shell as root to install the package, "unset PATH_PY311" will only
unset it in this shell, not the shell building the package.

Also fix a grammar issue in rustc, and unset LIBSSH2_SYS_USE_PKG_CONFIG
after installing rustc too.