zsh: Clean up documentation build and install instruction
The so-called "optional documentation" actually contains nothing more
than the HTML documentations generated by makeinfo --html.  So just stop
downloading it.

A change in texinfo-7.0 has actually break makeinfo --html of zsh, but
the issue was hidden due to the inclusion of the "optional
documentation".  Use a sed to correct it.

Do not put zsh.txt and zsh.html info the Doc directory, or they'll be
installed to /usr/share/doc/zsh-5.9/html by make install.html.

Also install the FAQ of zsh.