geoclue2: Raise libnotify to recommended
Though it's only for a demo, our policy is the package should build with
the unchanged book instruction with optional dependencies omitted.  So
this should be recommended.

Or we can add -Ddemo-agent=false into the command line but I think this
would be more nasty.

BTW sort the dependencies in alphabetical order.