texlive: No longer edit ld.so.conf
The texlive binaries are built with RPATH, thus they work fine w/o
ld.so.conf modification.

For other packages using kpathsea we need to symlink libkpathsea.so.6 to
/usr/lib so it can be found.

It seems install-tl-unx does not use texlive shared libraries at all (I
downloaded the tex executable from it and the executable only uses
libc.so.6 and libm.so.6).