Fix slight dissymmetry of Up/home positioning
The padding for the containing box was not right/left symmetric because
the order of padding values is not trivial in css (top right bottom
left). Making right and left equal allows to center the Up/Home box more
precisely. Then some tweaks for Next and Prev boxes have to be removed.