Make box for section titles static again
Since the  box that may scroll does not contain that box now, it does
not move when the scrollbar is actioned. The only fixed window now is
the div containing sect2/text. Its top has to be adjusted by trial and
error. But the rendering seems ok for all borwsers I can try (epiphany,
falkon and firefox).