nspr: Make sed more clear and remove unneeded "-r"
Quote from Rainer Fiebig via blfs-dev:

<quote>
I think that the "-r" in the first of the following seds is not
necessary and so may be a bit confusing as the next one uses just a
plain "-i":

sed -ri '/^RELEASE/s/^/#/' pr/src/misc/Makefile.in &&
sed -i 's#$(LIBRARY) ##'   config/rules.mk         &&

Personally, I also find the use of "#" in the second sed somewhat
unfortunate, as it used in the first one with a totally different
meaning.
</quote>

So let's remove -r and solve the "one symbol for different meanings"
issue.  Note that in '/^RELEASE/s/^/#/' the first two '/'s have a
different meaning than others too, so use "|" instead of "/" for the
separator in sed "s" commands.

Suggested-by: Rainer Fiebig <jrf@mailbox.org>