postfix - update the smuggling details for 3.8.5.
The previous 'yes' is now aliased to 'normalize' which will be
the default in postfix-3.9, so prefer that.  It now only rejects
if the bare newlines are in the DATA section of the mail, to
provide better compatibility with existing SMTP clients.

If a site needs to reject mails with bare newlines in the
message body, the value can be changed to 'reject'.

See https://www.openwall.com/lists/oss-security/2024/01/22/1