Preliminry update to kf6-6.0.0.
This update replaces kf5.  In doing so there are several packages
in BLFS that are broken, notably the LXQt packages that reference
kf5 packages.

Trying to maintain kf5 and kf6 in the book is more effort than it
is worth.  For those that want the kf5 based elements, blfs-12.1
still works.

Several packages were updated to build with qt6 so they are not available
for packages that need the qt5 versions,  These include qca, polkit-qt, and
phonon.  A couple of packages that are no longer needed for kf6 are still
in the book: kuserfeedback and libdbusmenuqt.

Many filenames still need to be changed, e.g. kf5-frameworks.xml, as
well as xml defined identifiers.

Later when LXQt releases their qt6/kf6 based packages, then we will
do a scrub of all the book internals to make names consistent.