Update most of kde-gear
There are still a few packages that need testing and the book updated.
The kf6 adn plasma packages referenced in LXQt are incorrect and need
to be changed to reference kf5 packages for now.

Note that there is a new package, plasma-activities, that is needed in
the kde-gear section, but needed to be taken from the plasma6 set of
packages.

kio-extras is no longer needed for the packages in the book.