texlive - remove context mkiv.
Explain why context is not present.

In the the unlikely event that somebody wishes to build context
from source, link to my old comments in the 2023 ticket, and to
the source tags at github.

Also mention that first installing (only) context from the binary
sounds like the easiest way to do this - the way we used to use the
full binary to bootstrap the source build.

Change a lingering 'command' to 'parameter' in the explanations.